# Performance optimization: 33.2% improvement in Dota 2 replay parsing