--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,4 @@
 /replays/*.dem*
 /tmp
 /vendor
+/cmd/manta-concurrent-demo/manta-concurrent-demo