--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@
 /vendor
 /cmd/manta-concurrent-demo/manta-concurrent-demo
 cpu.prof
+mem.prof