--- a/.gitignore
+++ b/.gitignore
@@ -4,5 +4,4 @@
 /tmp
 /vendor
 /cmd/manta-concurrent-demo/manta-concurrent-demo
-cpu.prof
-mem.prof
+*.prof