--- a/cmd/manta-concurrent-demo/go.mod
+++ b/cmd/manta-concurrent-demo/go.mod
@@ -1,0 +1,14 @@
+module manta-concurrent-demo
+
+go 1.21
+
+require github.com/dotabuff/manta v0.0.0
+
+replace github.com/dotabuff/manta => ../..
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/golang/snappy v0.0.3 // indirect
+	google.golang.org/protobuf v1.26.0 // indirect
+)