--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/dotabuff/manta
 
-go 1.16
+go 1.21
 
 require (
 	github.com/davecgh/go-spew v1.1.0
@@ -8,4 +8,9 @@ require (
 	github.com/golang/snappy v0.0.3
 	github.com/stretchr/testify v1.5.1
 	google.golang.org/protobuf v1.26.0
+)
+
+require (
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )