… .claude/settings.local.json
… .github/workflows/ci.yml
… .tool-versions
… CLAUDE.md
… ROADMAP.md
… cmd/manta-concurrent-demo/README.md
… cmd/manta-concurrent-demo/concurrent_benchmark_test.go
… cmd/manta-concurrent-demo/concurrent_test.go
… cmd/manta-concurrent-demo/go.sum
… field_state.go
… parser.go
… projects/2025-05-23-perf.md
… reader.go
… sendtable.go
… serializer.go
… stream.go
… string_table.go